% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_cool_table}
\alias{make_cool_table}
\title{a function to make a table using gt}
\usage{
make_cool_table(
  data,
  title,
  merge_cols = NULL,
  images_col_name = NULL,
  container_size = 450
)
}
\arguments{
\item{data}{the data set as a tibble}

\item{title}{the title of the table}

\item{merge_cols}{a vector of two column names in the table to be merged}

\item{images_col_name}{the name of the column that holds the path to the images}

\item{container_size}{numeric for the size of the html container}
}
\description{
a function to make a table using gt
}
