% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{show_me_a_logo}
\alias{show_me_a_logo}
\title{A function to show me a logo for a university
The function returns a path to the system directory location of the specific picture}
\usage{
show_me_a_logo(image, what_are_all_the_names = FALSE)
}
\arguments{
\item{image}{a single image name}

\item{what_are_all_the_names}{a Logical that will return all known logos in the package}
}
\description{
A function to show me a logo for a university
The function returns a path to the system directory location of the specific picture
}
